refactor gogle and adp
