+++ src/teamster/kipptaf/adp/workforce_manager/config/wfm-assets-daily.yaml
