+++ src/teamster/kipptaf/adp/workforce_manager/config/wfm-assets-dynamic.yaml
