+++ src/teamster/kipptaf/definitions.py
@@ -6,10 +6,8 @@ from dagster_gcp import BigQueryResource, GCSPickleIOManager, GCSResource
 from dagster_k8s import k8s_job_executor
 
 from teamster import GCS_PROJECT_NAME
-from teamster.core.adp.resources import (
-    AdpWorkforceManagerResource,
-    AdpWorkforceNowResource,
-)
+from teamster.core.adp.workforce_manager.resources import AdpWorkforceManagerResource
+from teamster.core.adp.workforce_now.resources import AdpWorkforceNowResource
 from teamster.core.alchemer.resources import AlchemerResource
 from teamster.core.amplify.resources import MClassResource
 from teamster.core.google.directory.resources import GoogleDirectoryResource
