--- src/teamster/kipptaf/google/config/assets.yaml
@@ -1,130 +1,0 @@
-assets:
-  - name: reporting__terms
-    range_name: src_reporting__terms
-    sheet_id: 1azcq9FsGDjYpvK7VBIHtGOsY8Yd-E5hFrxWuk5hFLH0
-  - name: surveys__staff_info_certification
-    range_name: src_surveys__staff_info_certification
-    sheet_id: 13bajVOf0otCUZCyLYk3KaTAr2yoOZ2IZGedGYIGwN8Q
-  - name: coupa__address_name_crosswalk
-    range_name: src_coupa__address_name_crosswalk
-    sheet_id: 1j7sBEH-PrkHTv83y3pYrgOPQvRm7S6vdXESbR_OYNFE
-  - name: coupa__school_name_crosswalk
-    range_name: src_coupa__school_crosswalk
-    sheet_id: 1j7sBEH-PrkHTv83y3pYrgOPQvRm7S6vdXESbR_OYNFE
-  - name: coupa__school_name_lookup
-    range_name: src_coupa__school_name_lookup
-    sheet_id: 1j7sBEH-PrkHTv83y3pYrgOPQvRm7S6vdXESbR_OYNFE
-  - name: coupa__user_exceptions
-    range_name: src_coupa__user_exceptions
-    sheet_id: 1j7sBEH-PrkHTv83y3pYrgOPQvRm7S6vdXESbR_OYNFE
-  - name: egencia__travel_managers
-    range_name: src_egencia__travel_managers
-    sheet_id: 1jMnltTr0cqJ5c1tlfyzySZ87eNaU6frCOD7E8f7FUc8
-  - name: egencia__traveler_groups
-    range_name: src_egencia__traveler_groups
-    sheet_id: 1jMnltTr0cqJ5c1tlfyzySZ87eNaU6frCOD7E8f7FUc8
-  - name: finance__enrollment_targets
-    range_name: src_finance__enrollment_targets
-    sheet_id: 1yCW2pmQYwLQuTel72QWCyq83qlyQEemYbONLPAIvy7c
-  - name: people__campus_crosswalk
-    range_name: src_people__campus_crosswalk
-    sheet_id: 1FCc28XWxFj3gSfItGGJ2tVU0C1fYD1JxKRxuSFqisMo
-  - name: people__employee_numbers_archive
-    range_name: src_people__employee_numbers_archive
-    sheet_id: 1FCc28XWxFj3gSfItGGJ2tVU0C1fYD1JxKRxuSFqisMo
-  - name: people__location_crosswalk
-    range_name: src_people__location_crosswalk
-    sheet_id: 1FCc28XWxFj3gSfItGGJ2tVU0C1fYD1JxKRxuSFqisMo
-  - name: people__powerschool_crosswalk
-    range_name: src_people__powerschool_crosswalk
-    sheet_id: 1FCc28XWxFj3gSfItGGJ2tVU0C1fYD1JxKRxuSFqisMo
-  - name: people__id_crosswalk_adp
-    range_name: fivetran_id_crosswalk_adp
-    sheet_id: 1xSa3dznVaGeqjo3Y0tS9GzkhVrpeCQ0aaKWlgI3kHik
-  - name: reporting__reporting_term_map
-    range_name: fivetran_reporting_term_map
-    sheet_id: 1azcq9FsGDjYpvK7VBIHtGOsY8Yd-E5hFrxWuk5hFLH0
-  - name: pm__teacher_goals
-    range_name: fivetran_teacher_goals
-    sheet_id: 1yepMCMDEQfYu-VwlzbtpSk5yCLfB76wtNFKiz_l180M
-  - name: pm__teacher_goals_term_map
-    range_name: fivetran_teacher_goals_term_map
-    sheet_id: 1yepMCMDEQfYu-VwlzbtpSk5yCLfB76wtNFKiz_l180M
-  - name: lit__gleq
-    range_name: fivetran_lit_gleq
-    sheet_id: 1yiCP5pTnOplhRw4u4wZeg1hW7GAnI2Hci5yecO1qLow
-  - name: renaissance__ar_individualized_goals
-    range_name: fivetran_ar_individualized_goals
-    sheet_id: 1UXyQhjJbyge05ydZjWDY6KnbuqVOw-PZpnmbY_ic57Q
-  - name: act__scale_score_key
-    range_name: fivetran_scale_score_key
-    sheet_id: 1QeVG9COJCTlRSPwjM9OQH7ofsAJcs7RxHe3WIE7nhW4
-  - name: lit__network_goals
-    range_name: fivetran_lit_network_goals
-    sheet_id: 1yiCP5pTnOplhRw4u4wZeg1hW7GAnI2Hci5yecO1qLow
-  - name: assessments__normed_scopes
-    range_name: fivetran_normed_scopes
-    sheet_id: 1G2z9rwXsFaMdFL6iOYdfQTVjZ7bctXMyz_Q09IhP4QE
-  - name: assessments__normed_subjects
-    range_name: fivetran_normed_subjects
-    sheet_id: 1G2z9rwXsFaMdFL6iOYdfQTVjZ7bctXMyz_Q09IhP4QE
-  - name: surveys__surveygizmo_abnormal_respondents
-    range_name: fivetran_surveygizmo_abnormal_respondents
-    sheet_id: 15PIpnnKaU2NF42JzjUbnrocL_2F41JOVZP4dYDS7mME
-  - name: ops__income_form_data
-    range_name: fivetran_ops_income_form_data
-    sheet_id: 1NH6njTDdxrTRhCxz7G19iDsMD7aBQSH1d20PDr6iRyQ
-  - name: assessments__fsa_iready_crosswalk
-    range_name: fivetran_fsa_iready
-    sheet_id: 1oBT61hVQVTS8QsiMqXaxH5IgPEw9xFdac6z0t_iygj8
-  - name: renaissance__ar_default_goals
-    range_name: fivetran_ar_default_goals
-    sheet_id: 1jC-1FvW2JOfsu1NbCCrcUtVkxkkfLVRYjEkpcpGp9ZQ
-  - name: payroll__register_code_lookup
-    range_name: fivetran_payroll_register_code_lookup
-    sheet_id: 1XUuuBhjh0fyKMEw-iW1eEGg0usaYR3LdHc-4sExvlpQ
-  - name: pm__assignments
-    range_name: fivetran_pm_appsheet_assignments
-    sheet_id: 1LMuwh41lIwdbEa_lQ9p1kzJmGIOTBTFz12Scb6_TQMM
-  - name: people__nj_cert_tracker
-    range_name: nj_cert_export
-    sheet_id: 1HNe3V4yABln436GMVJyiHnjluU-xCXS1Oq289gg7PnM
-  - name: reporting__school_health_metric_lookup
-    range_name: fivetran_shd_metric_lookup
-    sheet_id: 10iUuu4HhLIem1WQQH5G4Y1V4khBhfAxRpoD6EClr7Sw
-  - name: surveys__scds_responses
-    range_name: fivetran_surveys_scds
-    sheet_id: 1m_OUhx7XMkW7rMMl11z19eN-5Z93fGwKqUKknjRc_K0
-  - name: assessments__power_standards
-    range_name: fivetran_power_standards
-    sheet_id: 1x68_1ZH5vaQTo4ynTHqAxx2R3SDTLL-N7czsQ_rlDjI
-  - name: assessments__standards_translation
-    range_name: standards_translation
-    sheet_id: 1EwHv7MajgHTq-HphLpwrS0KG_2Y_FXlPuPvj98YuFS8
-  # - name: njdoe__cert_export
-  #   range_name: nj_cert_export
-  #   sheet_id: 1HNe3V4yABln436GMVJyiHnjluU-xCXS1Oq289gg7PnM
-  # - name: people__staff_attendance
-  #   range_name: fivetran_staff_attendance
-  #   sheet_id: 16AB2wcagWOExCo_wjNfmxD2fcxoTwJHWVyrn77dX6VA
-  # - name: people__staffing_model
-  #   range_name: fivetran_staffing_model
-  #   sheet_id: 1KzRMjL2kmC4xV3gllh70g8HReVhjvHrqNHEl0WCVM24
-  # - name: people__student_password_override
-  #   range_name: fivetran_student_password_override
-  #   sheet_id: 1nHUip20GCnE9EFbbuUz-Ho8HvKvb126Qn4BHOc0eSCg
-  # - name: pm__pir_rubric_scores
-  #   range_name: fivetran_pir_rubric_scores
-  #   sheet_id: 1SyTRcMKmLjVLzPfgPU8lZ8fat86f6e0Rr_s4gZMVmPI
-  # - name: pm__teacher_goals_exemption
-  #   range_name: fivetran_teacher_goals_exemptions
-  #   sheet_id: 1d84x7HHOMgnQM78dYLY9vHqf-jLfYmpH8YGHEDezGNs
-  # - name: reporting__report_card_comment_bank
-  #   range_name: fivetran_report_card_comment_bank
-  #   sheet_id: 1azcq9FsGDjYpvK7VBIHtGOsY8Yd-E5hFrxWuk5hFLH0
-  # - name: surveys__exit_survey
-  #   range_name: fivetran_exit_survey
-  #   sheet_id: 1ONA_w4YYl7fSSPsUkx5H5mqtvyXqvBTGMrgnHIL73jA
-  # - name: surveys__so_assignments
-  #   range_name: fivetran_so_assignments
-  #   sheet_id: 15PIpnnKaU2NF42JzjUbnrocL_2F41JOVZP4dYDS7mME
