--- src/teamster/kipptaf/google/schedules.py
@@ -1,51 +1,0 @@
-from dagster import RunRequest, ScheduleDefinition, schedule
-
-from .. import CODE_LOCATION, LOCAL_TIMEZONE
-from .assets import FORM_IDS
-from .jobs import (
-    google_directory_nonpartitioned_asset_job,
-    google_directory_role_assignments_sync_job,
-    google_directory_user_sync_job,
-    google_forms_asset_job,
-)
-
-google_directory_nonpartitioned_asset_schedule = ScheduleDefinition(
-    job=google_directory_nonpartitioned_asset_job,
-    cron_schedule="0 1 * * *",
-    execution_timezone=LOCAL_TIMEZONE.name,
-)
-
-google_directory_role_assignments_sync_schedule = ScheduleDefinition(
-    job=google_directory_role_assignments_sync_job,
-    cron_schedule="30 2 * * *",
-    execution_timezone=LOCAL_TIMEZONE.name,
-)
-
-google_directory_user_sync_schedule = ScheduleDefinition(
-    job=google_directory_user_sync_job,
-    cron_schedule="0 0 * * *",
-    execution_timezone=LOCAL_TIMEZONE.name,
-)
-
-
-@schedule(
-    cron_schedule="0 0 * * *",
-    job=google_forms_asset_job,
-    execution_timezone=LOCAL_TIMEZONE.name,
-)
-def google_forms_asset_job_schedule():
-    for form_id in FORM_IDS:
-        yield RunRequest(
-            run_key=(
-                CODE_LOCATION + "_google_forms_static_partition_assets_job_" + form_id
-            ),
-            partition_key=form_id,
-        )
-
-
-__all__ = [
-    google_directory_nonpartitioned_asset_schedule,
-    google_directory_role_assignments_sync_schedule,
-    google_directory_user_sync_schedule,
-    google_forms_asset_job_schedule,
-]
