+++ src/teamster/kipptaf/google/sheets/sensors.py
@@ -1,6 +1,6 @@
 from teamster.core.google.sheets.sensors import build_gsheet_sensor
 
-from .. import CODE_LOCATION
+from ... import CODE_LOCATION
 from .assets import google_sheets_assets
 
 google_sheets_sensor = build_gsheet_sensor(
@@ -8,7 +8,3 @@ google_sheets_sensor = build_gsheet_sensor(
     asset_defs=google_sheets_assets,
     minimum_interval_seconds=(60 * 10),
 )
-
-__all__ = [
-    google_sheets_sensor,
-]
